Port the NDXR container walk, and fix a control I had weakened in translation
Adds include/ac6/retail_ndxr_container.h and src/retail_ndxr_container.cpp: the
container walk - recognise, GIDX unwrap, type code, four section extents, the
fixed 0x30 record array, the string table - with the header citing the address
that establishes each stage. It stops at the polygon descriptor deliberately,
because cycle 1203 showed the product's reader and this derivation disagree there
and neither mapping is controlled; porting one would swap a measured guess for an
uncontrolled one. Two deviations from retail are written into the header rather
than assumed: this reader does not relocate the buffer, and it checks [+0x04]
against the file length, which retail does not.

The new test failed on first run and the fault was mine. It asserts cycle 1196's
discriminating control - the string table is at [+0x10] + 0x30 and the rival
without it must fail - which in Python read sixteen bytes and scored 0 of 537. In
C++ I wrote it as a NUL-terminated printable run, which one stray character
satisfies, and the rival passed in 78 of 537. Measured across thresholds: at 1
the rival scores 78, at 2 it scores 30, at 4 and 8 it scores 0 while the derived
base still names all 537, and at 12 the derived base drops to 513. The threshold
is 8, inside that plateau, chosen from the table rather than raised until green,
and the table is in the test's comment.

The lesson is narrower than test-your-tests: a control re-expressed in another
language is a NEW control and inherits none of the first one's evidence. It is a
tenth shape for INSTRUMENT_DISCIPLINE - a weakened control that still passes on
the true hypothesis and so looks healthy from the green side.

Google C++ Style Guide applied to the three new files as instructed: no macros,
CamelCase for non-accessor functions, accessors left snake_case as the guide
permits, 80 columns, kConstantCase. This diverges from the rest of the product,
which is uniformly snake_case; the sweep is owed and is deliberately not ridden
along with a derivation commit.

537 files, 13,014 records, 0 refused, 0 unnamed, 0 with the guard set, 537/537
named from the derived base and 0 from the rival. ctest 27 tests all passed,
JF=pass.

+++ reconstruction/ace-combat-6/src/native_geometry_raster_target.cpp
@@ -595,8 +595,10 @@ void NativeRenderTarget::rasterize_clipped_triangle(GeometryRasterContext& conte
     const float safe_x = std::clamp(vertex.x, -1.0f, 1.0f);
     const float safe_y = std::clamp(vertex.y, -1.0f, 1.0f);
     return ScreenPoint{
-        std::min(width_ - 1u, static_cast<std::uint32_t>((safe_x * 0.5f + 0.5f) * width_)),
-        std::min(height_ - 1u, static_cast<std::uint32_t>((0.5f - safe_y * 0.5f) * height_)),
+        std::min(width_ - 1u, static_cast<std::uint32_t>(
+                                  (safe_x * 0.5f + 0.5f) * static_cast<float>(width_))),
+        std::min(height_ - 1u, static_cast<std::uint32_t>(
+                                   (0.5f - safe_y * 0.5f) * static_cast<float>(height_))),
         std::clamp(vertex.z * 0.5f + 0.5f, 0.0f, 1.0f),
         vertex.x, vertex.y, vertex.z};
   };
